PDF Generation fixed and period selection enabled
Code needed a lot of
refactoring since the last version of the system had some time issues due
to a bad SQL query that took so long because a missunderstanding
of how JOIN clauses works.

On this new version python is taking care of the filtering, improving
a lot the speed of PDF generation, also the a new option for
report by period generation was added.